dfstate: Add `persistent_state::handle` mod
Adds a new `persistent_state::handle` module that houses
`PersistentStateHandle`. In a future commit, we will place the RocksDB
handle and the shared state under separate `RwLock`s, which opens up the
potential for deadlocks. To remove the possibility of deadlocks, we'll
rely on the fact that `PersistentStateHandle`'s struct members are
private to the `persistent_state` module, which will require users of
`PersistentStateHandle` to interact with the locked data solely through
its public methods. These methods will be written in a way that makes it
impossible for deadlocks to occur by ensuring that the locking of the
RocksDB handle *always* occurs before the locking of the shared state.

This is all necessary because we'll need to be able to lock the RocksDB
handle separately from the shared state in order to sync the WAL to disk
without blocking other writes to the table.

Refs: REA-3434
